New biome, plus new bush type. (#38)
* Update pack.yml

Might as well add my name to this.

* Update colors.yml

Highlands biome colour.

* New "highlands" biome.

Largely based off the code for temperate_mountains, so this is mostly just my attempt at trying out the creation of a biome. Needs polish.

* Highlands biome added to distribution.

Current value is temporary, just using it for testing purposes. I primarily just want to break up the temperate mountains.

* Add spruce_bush_patches.

Used for "highlands" biome. Hopefully this works!

* Fix weird grass issue, plus change bush type.

...hopefully, at least. >.>

* Revert to old bush type.

* Changed highlands colour. Should be better!

* Remove moss_stone generation.

* Bump version, remove name.

* Lower frequency of highlands.

* Revert unintentional change.

Not sure when this was flipped, but uh, woops.